Run git commands from top-level
The plugin’s usage of “git ls-tree . . . <file>” and “git log . . .
<file>” assumes that the commands are being run from the top level of
the working tree. Ensure that this is the case by explicitly changing
the working directory to fugitive_repo.tree() while issuing these
commands.

This fixes :GV! when it is used with a current working directory other
than the top-level.