Checkout feature
I love using Git Graph and how easy is to visualize the Git Tree and play around with branches and commits. I think a great addition to gv features could be a mapping to checkout a specific commit. So, this is my naive implementation, feel free to give other ideas of how make this betters.