Fix windows split: do not open a new window even if gv window was moved
Problem: When reusing the splitted window to show a commit (that has
`w:gv` = 1), only the rightmost window (winnr = '$') is considered.
This results in a weird behavior that, if the window is moved to top
or left, another window will be opened in split.

Solution: Search all the windows in the current webpage to find the
target window with `w:gv = 1`.